+++ examples/sql_crud_simple.py
@@ -61,8 +61,8 @@ TableModel.Attributes.sqla_metadata.bind = db
 
 
 class Permission(TableModel):
-    __tablename__ = 'spyne_user_permission'
-    __namespace__ = 'spyne.examples.user_manager'
+    __tablename__ = 'permission'
+    __namespace__ = 'spyne.examples.sql_crud'
     __table_args__ = {"sqlite_autoincrement": True}
 
     id = Integer32(primary_key=True)
@@ -71,8 +71,8 @@ class Permission(TableModel):
 
 
 class User(TableModel):
-    __tablename__ = 'spyne_user'
-    __namespace__ = 'spyne.examples.user_manager'
+    __tablename__ = 'user'
+    __namespace__ = 'spyne.examples.sql_crud'
     __table_args__ = {"sqlite_autoincrement": True}
 
     id = Integer32(primary_key=True)
@@ -85,18 +85,31 @@ class User(TableModel):
 @memoize
 def TCrudService(T, T_name):
     class CrudService(ServiceBase):
-        @rpc(Integer, _returns=T)
+        @rpc(Mandatory.UnsignedInteger32, _returns=T)
         def get(ctx, obj_id):
             return ctx.udc.session.query(T).filter_by(id=obj_id).one()
 
-        @rpc(T, _returns=Integer)
+        @rpc(T, _returns=UnsignedInteger32)
         def put(ctx, obj):
-            ctx.udc.session.merge(obj)
-            ctx.udc.session.flush()
+            if obj.id is None:
+                ctx.udc.session.add(obj)
+                ctx.udc.session.flush()
+
+            else:
+                if ctx.udc.session.query(T).get(obj.id) is None:
+                    raise ResourceNotFoundError('%s.id=%d' % (T_name, obj.id))
+
+                else:
+                    ctx.udc.session.merge(obj)
+
             return obj.id
 
-        @rpc(Integer)
-        def del_user(ctx, obj_id):
+        @rpc(Mandatory.UnsignedInteger32)
+        def del(ctx, obj_id):
+            count = ctx.udc.session.query(T).filter_by(id=obj_id).count()
+            if count == 0:
+                raise ResourceNotFoundError(obj_id)
+
             ctx.udc.session.query(T).filter_by(id=obj_id).delete()
 
         @rpc(_returns=Iterable(T))
